move nations stuff
