+++ Nations/src/main/java/net/pixlies/nations/entities/User.java
@@ -1,0 +1,5 @@
+package net.pixlies.nations.entities;
+
+public class User {
+    // TODO: FPlayer
+}
