+++ Nations/src/main/java/net/pixlies/nations/nations/NationsEntity.java
@@ -1,4 +1,4 @@
-package net.pixlies.nations.entities;
+package net.pixlies.nations.nations;
 
 public interface NationsEntity {
     String id();
